[iwlwifi] Synchronize usage of mvmvif in rx ISR
Unsynchronized usage of mvm->mvmvif[0] in the receive ISR in rxmq.c
led to driver crashes because mvm->mvmvif[0] could be deleted or in an
unusable state due to writes to the object from a separate thread.

This commit adds RCU synchronization and checks to ensure that
mvm->mvmvif[0] is in a valid state before using it in the receive ISR.

Bug: fxb/97275, fxb/98324
Test: Create and delete wlan ifaces 50 times using wlan-dev with
intervals of 0.2,  1, 5, and 15 seconds between creation/deletion,
and observe no driver or firmware crashes. This test was repeated
twice: once without associating with an AP, and once while a YouTube
video was playing in the background after associating with an AP.

Change-Id: Ief9881867dc5dbf4c11712cff12ecdc834f1f882
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710223
Reviewed-by: Louis Lo <yjlou@google.com>
Commit-Queue: John Tan <johntan@google.com>