[e2e] patch the ermine terminal e2e test for CFv2 monikers
This is a simple fix so the e2e test does not break because it is using
CFv1 monikers. The CFv2 monikers for ermine elements have a randomized
component name, so the complete path is not available to the test, but a
wildcard pattern should work, as the test is currently implemented.

(We are investigating alternative approaches for a future patch.)

Bug: 106471

Change-Id: I74fa292d1148cd167f83b0a61ced0efaf4622c49
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/711334
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>