[pigweed] Configure pw_log backend for DFv1
Create a pw_log backend that uses DFv1.

Make DFv1 the default pw_log backend. The only place pw_log
is needed right now is in a driver.

Reorganize Pigweed backends so that that are under a directory
with the name of the module that contains the frontends.

Bug: 105192

Change-Id: I3cee607bd7a6a4381f309ff8cbb397be18146205
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704150
Reviewed-by: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Ben Lawson <benlawson@google.com>