[fshost] Add pause and resume to block watcher
This is used by several integration tests to set up block devices for
fshost without fshost latching onto them before we are done, so it's
useful to get in early.

This also disables a flakey test. The issue is resolved via a more
complex refactor in fxrev.dev/708822 but that isn't going to go in
immediately and it has a high flake rate that is preventing this cl from
going in (and causing other issues).

Bug: 87511
Bug: 106137
Test: fx test fshost_rust_tests
Change-Id: Ibf3ac11fb1ebaa2cdcf7f45a9d8bd3ea94ae62e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701683
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>