[check-licenses] Update metrics template for files
Every package in check-licenses implements the metrics template.

At the end of the run, these template methods are used to collect all
the metrics and results of the run and save them to disk.

With the new feature request of fxbug.dev/104400, I now need to save
the full license segments to disk at the end of the run.

This CL adds a "files" entry to the template instances for all the
packages in check-licenses. Future CLs will actually use this new field.

Bug: 104400

Change-Id: If68712b05b2e98d029b62f375a919a298eb42553
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707990
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>