[fdf] Add matched_composite into MatchedDeviceGrp
Adds the matching composite info into the matched device group when
returning it to the client. Store the transform and the matched
composite info in a hashmap that maps the device group's topological
path to the two.

Bug: 105313
Tests: Added unit test for this to the driver index tests
Change-Id: Iecf7d000bad3be64615f1696050191e1a872a4c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707322
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>