[ffx][daemon] Graceful shutdown of main loop
Currently the daemon exits after a 20ms timeout from getting a quit
message, and it does so by simply calling exit(), which is likely one of
the reasons it winds up leaving files behind in inconsistent or weird
states.

This change makes it so that the quit message triggers an mpsc message
back up the loop to cause it to end normally, allowing the caller (and
any other threads running) to clean up after itself. Also removes an
unnecessary panic! in the event of an error processing a message, also
using the quit message to cause the process to exit.

Bug: 76362
Bug: 106578
Bug: 82518

Change-Id: I26b1756fa36055606b3a10143318becd293daaf5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711544
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>