[a11y] ScreenReaderMessageGenerator: use "out parameters" in Describe helpers
Context: there are a bunch of helper functions in
`ScreenReaderMessageGenerator` that are called as part of `DescribeNode`
to stitch together a description of a semantic node.

This CL changes all such helper functions that had previously returned a
vector of utterances; they now append to an "out parameter" vector
instead.

This is purely a mechanical change. There's no logic changes at all.

This should reduce boilerplate and make it easier to refactor the logic
in this file more in the future.

Bug: 81707
Test: fx test a11y-tests a11y-lib-tests
Change-Id: Ic8289475209f67b4f8796f729d22f73f682c2099
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709617
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Fuchsia-Auto-Submit: Kevan Hollbach <khollbach@google.com>