[check-licenses] Add license texts and patterns to Files map.
This CL adds relative paths to the File and Pattern structs,
so they can be added to the metrics Files map using unique keys.

In a future CL, the "result" package will be updated to write the
contents of the Files map to disk at the end of the run.

Bug: 104400
Change-Id: I607804d5b2f5ffdfeadc05e78e9e765f54aa6001
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709242
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>