[scrutiny] Extract structured configuration to a file in the build.
Can be used to regularly check for the addition of new configuration
keys in a system image.

Adds a new integration test which transitively depends on the system
image assembly, so we wire it up using the :host_tests pattern we're
using for assembly tests whose deps can't be included in the graph
of the system image.

Bug: 105366
Change-Id: I516869315c330babd0e26c91643a3fa9db1666eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705598
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>