[component][fuzzing] Replace ReadFromCorpus with GetCorpus
Early on a decision was made to return inputs from a runner's corpora
one at a time. The exact reason for this decision has been lost to time,
but it led to corpus retrieval being somewhat brittle, and, in the case
of libFuzzer, led to a bug where the first input could not be retrieved.

In practice, callers of `ReadFromCorpus` only ever wanted the whole
corpus, so this CL replaces that method, which took an offset and
returned a single `Input`, with `GetCorpus`, which returns a vector of
all non-empty inputs.

Change-Id: Ie0b8be86f914897970de4e9e5c7e705deae005ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711504
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>