[fidl][cpp] OutgoingToIncomingMessage requires non-transactional messages
In the next CL we will introduce a non-transactional version of
|IncomingMessage|, a class managing the body portions of FIDL
transactional messages. Non-transactional uses of |IncomingMessage|
would move to that class. Because of that, we have
to make a choice whether the |OutgoingToIncomingMessage| utility
returns the transactional message class or the non-transactional one.

Because more production uses of that utility are non-transactional, this
CL makes that a requirement.

Bug: 82681

Change-Id: I680b5c4106f4bf6947fb68d48bb18a68d8b34bb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709463
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>