[camera] Fix another uninitialized memory access.
When the camera is disabled during boot, Imx227Device::current_mode_ is
left uninitialized when Imx227Device::CameraSensor2SetMode() errors out.
Later, the driver crashes when using the uninitialized current_mode_ to
index available_modes in Imx227Device::CameraSensor2GetExtensionValue().

Fix the crash by setting an initial value for current_mode_. Setting an
initial value is necessary, rather than returning errors for queries
before a mode is set successfully, because the setup of InputNode /
ProcessingNode does not gracefully handle errors and asserts if any
queries fail. This mitigation is not ideal, but it prevents a crash that
otherwise results in the camera subsystem failing to initialize.

Bug: b/228516004
Bug: b/242083127
Test: Video functions after booting with camera hw switch off.
Change-Id: Ic17bbf82bef6efcc6bab40d6c98cd90de443416d