[zither][go] Introduce go data layout bindings
Partially as a demonstration of the utility of the machinery we've
built, we add a go backend to zither with minimal effort. It will serve
to have something not entirely C-like as a reference in further
development, and this backend will later be needed to deprecate kazoo's
generation of bindings for the go runtime.

Bug: 91102
Change-Id: I65b10a3b54b9a8a7ec7c0f142a0d07621fd67c36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708763
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>