[pbms] Generate product bundle for bringup
This CL enables the generation of a product bundle when building
the bringup product. The schema is adjusted to allow for an empty image
list, since bringup does not generate an FVM, and the `ffx emu` tool is
likewise adjusted to accept an empty FVM list.

Change-Id: Icdaf77a9684263e6c8aee420c0cfcf6edc273d0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712545
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>