[fit] Reduce unnecessary padding in fit::function_base
The fit::function_base has two data members, in order:
 * ops_: a pointer to a struct and
 * bits_: a std::aligned_storage<inline_target_size> (assuming
   inline_target_size is of at least sizeof(void*)).

The issue with this arrangement is that in 32-bit architectures where
the alignment of max_align_t is 8, there's always a 4 byte padding
between the two fields, even if inline_target_size is also 4. The
alignment of the bits_ field is always as large or more than the
alignment of the ops_ pointer, so this patch reverses the order of the
fields to allow the compiler placing the ops_ pointer at an alignment
smaller than max_align_t if possible.

However, since the storage_type (type of bits_) internally handles the
alignment requirements for values of inline_target_size that are not
multiple of max_align_t, this patch also moves the alignment
requirements from the storage_type to the member field and the only
other usage of storage_type in the class.

For 32-bit systems, this reduces the size of fit::function_base<N> by 8
bytes when N is an odd multiple of 4, meaning that fit::function_base<4>
is now 8 bytes instead of 16 bytes.

Change-Id: I70771c6a991f662a80371b3e6158ece9b3fc58ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709609
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Alex Deymo <deymo@google.com>