[build][test] Fix package_name for fuchsia_test
This CL corrects a discrepancy between the `fuchsia_test_package` and
`fuchsia_test` GN templates when a `package_name` is provided. In the
former, the `target_name` is used to derive the `package_manifest`
before the check to see if `invoker.package_name` is defined. The latter
constructed `package_manifest` after this check and thus used the value
of `invoker.package_name` if defined. This mismatch caused tooling
designed to work with `fuchsia_test_package`, like `test_list_tool`, to
fail if `fuchsia_test` was given a `package_name`.

Change-Id: I96b551d5dcd610ddaf388e0ff0ab0cfc4f4b375a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712806
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>