[fidl][cpp] Natural types printing
This adds the ability to print natural objects. Because this could add
bloat to programs that don't need it you need to add an explicit build
dependency on libraryname_natural_ostream and include the
fidl/libraryname/cpp/natural_ostream.h header.

The printed format is not JSON or C++ or FIDL. It is not stable and will
change. It must not be used for tests.

Documentation will be coming in a follow-up CL.

Change-Id: I2cb21993c82a4a28affdb8d826aad80b5d4bfd31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/702513
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>