[kernel][irq] Check IRQ config when DEFAULT config is passed.
If a user attempts to create an interrupt and specifies "default" for
the trigger mode and level of the interrupt, be sure to check the
current configuration of that interrupt before configuring the
interrupt dispatcher.

Failure to do this means that we might have an interrupt which has
been configured for level triggering (because of ACPI), but then fail
to note that we need to mask and unmask the interrupt at the start/end
of processing.

Bug: 102058
Change-Id: Iadc658b5c3aafcf721d168ee7322fd85965b80e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712703
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>