[kernel][irq] Fix a bad range check in remap_interrupt
remap_interrupt on the PC platform was checking to see if the
interrupt vector being passed was in the range from [0, NUM_ISA_IRQS],
and passing the vector on to apic_io_isa_to_global if it was.  In
several places in ioapic.cc, however, it is made clear that the proper
range should be [0, NUM_ISA_IRQS).  There are several places where
the code ASSERTs that vector numbers are strictly < NUM_ISA_IRQS, and
the vector itself is frequently used as an index into statically sized
arrays of NUM_ISA_IRQS in length.

So, fix the check in remap_interrupt so that the range is correct and
does not accidentally trigger an assert.

Change-Id: I90dde6bc01156cbce52fc7296ac345b7b69eac98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712704
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>