[fit] Fix internal class template parameters
The function_impl, callable_impl, and function_base classes were
declared as if they had Result and Args... as template parameters, but
this was not the case. These classes are actually templated on the
callable type, and the result and argument types are extracted in the
partial specialization for function types. The Result parameter was
used for the callable type and the Args... parameter pack was always
empty.

Change-Id: Id1dd813c1a76efebd39c796c0b41e1bbaf83ef6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710573
Fuchsia-Auto-Submit: Wyatt Hepler <hepler@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>