[ntr] Add support for starting DHCPv6 client
Add support for starting DHCPv6 client when requested. Only one client
can be started on an interface at any given time. This limitation
shouldn't be an issue, as the RFC recommends against running multiple
clients on an interface, and so it's unlikely this will need to be
supported in the future.

Changed DHCPv6 client to block `WatchAddress` forever rather than
closing the channel and stopping itself so that it's possible to
exercise the watch stream.

Add ffx plugin test and an integration test which verifies that the
initial Solicit/InformationRequest message is sent by the DHCPv6 client
when started in stateful/stateless modes respectively.

Test: fx test --test-filter 'start_dhcpv6_client::*' \
              network-test-realm-integration-test

Change-Id: I872f221cddd36b426bc22025af7dd56b355096ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707509
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Nate Hurley <nahurley@google.com>