[f2fs] Fix a buffer overflow on SuperblockInfo::checkpoint_block_
This CL correctly calculates kMaxSitBitmapSize based on which the
NAT bitmap size is determined. Previously, kMaxSitBitmapSize was
always set to 1, and the NAT bitmap size was overestimated, which
would cause the buffer overflow on checkpoint_block_ and corrupt
other members of SuperblockInfo on a large-volume. It occured
on 1TB NVMe/NUC11TNHi7.

Change-Id: I9eec7bc41c1a1f5128d0915da85b392ac19f8fa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712882
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Seokhwan Kim <sukka.kim@samsung.com>