[emu][sdk] Adding aemu arm64 prebuilts to SDK
This adds the aemu_internal for arm64 atoms to the core SDK.

The prebuilts have been added to the manifest and roller updated in a CL
in the infra repo.

They are optional for a developer checkout, so to enable generating the
arm64 files in the SDK,
Run: `jiri init -fetch-optional=arm-sdk-tools` and `jiri update`.

Then include `--args arm-sdk-tools=true` in your `fx set`.

qemu arm64 is pending on fxbug.dev/106608, so will be added in a
separate CL.

Bug: 99321
Change-Id: I6eb7a3bf736b0b54129ae9dcf802413cef9d1677
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711264
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>