[scenic] Skip tests allocating non-dedicated external memory on FEMU
For FEMU and other devices that don't have UMA architectures, if clients
allocate non-dedicated exportable memory, the ICD cannot determine
the sysmem memory constraints and which sysmem heap to use merely based
on the memoryType, and the test will fail.

Currnetly the only usages of AllocateExportableMemory() are tests.
These tests need to be skipped on FEMU and we should migrate them to
import sysmem BufferCollections instead, which is the current preferred
way to create exportable images / buffers.

Bug: 107082

Test: gfx_unittests, scenic_pixeltests on ToT goldfish-opengl
Change-Id: I9256f1edadcd436babd5f1c7fbd5ef8fac291d5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711413
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Shivesh Ganju <shiveshganju@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>