[benchmarks] Remove the disabled touch_input_latency test
This was added in 2020 but was disabled not long after it was added.

It is one of the few remaining performance tests that refers to a
Components v1 component.  It refers to touch-input-test.cmx, which was
removed (and converted to touch-input-test.cml) in
https://fuchsia-review.googlesource.com/c/fuchsia/+/559102.

Bug: 86575
Fixed: 76494
Change-Id: I75c3c3ae5eb01b5d50f83109daf62ecc687087a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709816
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>