[overnet] Add new torture test to QUIC
This also simplifies some of the locking logic in a way that shouldn't
affect functionality but made the code easier to read about while
debugging.

It also handles an error condition when we bind a remote stream and
immediately attempt to write to it even though the other end hasn't
established the stream yet. The latter shouldn't affect much either;
producing a situation where this happens and behavior is correct is much
harder than I initially realized, and I ended up deleting the test case
that would have exercised it because it was not and could not be
correct, but the change and the comment are worth keeping for posterity.

Change-Id: I51f25870f264cd6468ecec72720ff65d943d82f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710662
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>