[ffx][pbms] add VirtualDeviceProduct
Add a wrapper for the product bundle data for a virtual device
to generalize the pbms API by abstracting out the schema versioning.

The emu tool will still have to understand and handle different
versions of the virtual device spec, but not the product bundle
metadata.

The pbms lib will understand and handle different versions of the
product bundle metadata, but not the virtual device spec versions.

I.e. This separates the concerns/domains properly.

Change-Id: Ice349ff6867693f1843c04e6d6d979006c37f451
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711561
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>