[testsharder] Don't skip tests if there are no affected tests.
A change for which we are unable to determine affected tests would tell
testsharder there are no affected tests and it would end up skipping all
hermetic tests as a result. Instead, we should assume that if we are at
the testsharder step, we've determined that we want to run tests (either
because there are affected tests or we want to run all tests), so if we
see there are no affected tests, we should just run all tests.

Bug: 107088
Change-Id: I9754971240c221d3a85db9a0b603c5a7357514bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712384
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>