[syslog] Do not inline LogNew
We can encounter a stack overflow in blobfs that prevents networking
from working if hwasan is enabled. The codegen caused 4 LogBuffers to
be added to the stack which would decrement SP by ~131kB. This
behavior can also be seen with asan without the stack-to-heap
promotion. This occurs from a combination of inlining, a large
LogBuffer, and sanitizers not allowing stack coloring to work
correctly. Until we can get hwasan/asan to work with stack coloring,
or the syslog API can be reworked to avoid adding a large buffer on
the stack, we can unblock the hwasan rollout by marking the function
that creates the LogBuffers as inline. This may likely be a small
penalty since this is usually done in a failure handling path.

Bug: 106574
Change-Id: I9eaa1da440271d22e07b10253e6390b8ddf19226
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713441
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>