[usb][xhci] Use deadline profiles for XHCI IRQ threads.
The XHCI threads responsible for handling USB interrupts (the
"completer" threads) were using high weight fair profiles, instead of
deadline profiles as they should have been.

Using a high weight fair thread is like saying that you need a lot of
throughput, but that latency is not so important to you.  A deadline
thread is what is needed when latency is the main issue.

Also, instead of hard coding the profile definition in the source
itself, move it core profile definitions and then apply it to our
thread using the profile name.  This will not only make it much easier
tune the profile if needed, but also to override it on a product by
product basis if needed.

Finally, do something similar for the USB audio callback thread.  It
was using a proper deadline profile, but the profile was hardcoded.
Now, it is defined by central config instead.

Bug: 102058
Change-Id: Icf297df4f67964225e33196468fbdfdba6c02fe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708523
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>