[system-updater] Fetch images from images.json if present
If an images.json manifest is included in the update,
fetch the images that need to be written. Be sure to
use the resource provided to find the image rather than the
current assumption of imagetype.name().

Update the retained index at each step to guarantee forward
progress.

Adds tests for parity with the soon-to-be-deprecated
update package format.

Fixed: 95804

Change-Id: I9e5a50a35c43fc7755581773b73050d2739a6cb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/704011
Commit-Queue: Olivia Nordquist <nolivia@google.com>
Reviewed-by: James Buckland <jbuckland@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>