[wlan] Pack BitField types
Clang has changed it's behavior in attempt to match gcc's
by not packing non-pod types unless those types are
explicitly packed also. In this case the BitField types
were not marked and are not pod. LittleEndianBitField can
be trivially made pod by making it's data members public.
BitField would need to not default initialize val_ to 0,
which is untenable. So instead the address-of-packed-member
warning is suppressed to mark these types as packed.

Marking these types packed will keep the same layout across
Clang abi versions. It was a bug in clang’s
address-of-packed-member warning that it wasn’t warning
before. ASan caught this UB which is why -fno-sanitize=alignment
is currently being used, this is a known problem in the code
so suppressing the warning is sensible.

Bug: 92521
Change-Id: I996e6d891e59c34982a6572679ac6ef53b368a7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713825
Reviewed-by: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>
Commit-Queue: Alex Brachet <abrachet@google.com>