[kernel][wait] Fix a wait queue ordering bug.
Fix a bug in the selection routine used for wait queues.  Basically,
there was a situation where if you had a wait queue with at least one
deadline thread in it, and at least one fair thread, and all of the
absolute deadlines of all of the threads in the queue had expired
relative to the query time, the Peek operation would return a fair
thread instead of the deadline thread with the minimum relative
deadline.

The fix for this is simple.  Also add some tests which check all of
the basic ordering properties of the collection.

Bug: 102058
Change-Id: I52c553cd6e735af828c6fe95ea5fc9e5d25145e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713514
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>