[ffx][config] Make env file handling async
Introducing lockfiles to the environment api introduced hard waits into
the normal config flows, which may be affecting stability in callers of
general config stuff. Change it to async so that we don't ever block the
process trying to load config beyond normal file loading/saving
blocking.

Also, panic if we try to set the config more than once outside of the
testenv infrastructure. This should have gone in a while ago, and the
INIT_ONCE had to go to make init() async properly so this is a good time
to correct that.

Multiply: ffx-e2e
Fixed: 107090
Change-Id: I9fb402bd3e752981f29ffb2e8762ae25a213cb1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712378
Reviewed-by: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>