[netstack3] Buffer creation for active open connections
The bindings need to pass down an object (zircon socket on Fuchsia) to
Core to create buffers later because `Describe` could be called before
calling `Connect` so the Bindings must have created the zircon socket
pair before initiating a connection. This change makes the state machine
be aware of the difference and tracks the object. The buffer creation
for the passive open path is unchanged. `ClientEndBuffer` doesn't make
much sense in the active open case, splitted it into two types:
`PassiveOpen` and `ActiveOpen` which are both `()` in this CL.

Change-Id: I9abcd3b41ddda85f58d6b4eb44675afc058fbab8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712589
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>