[ffx][config] Clean up bad lockfiles proactively.
Do a few simple heuristic checks to deal with the most common situations
where a lockfile might be left lying around and delete them when we can.

This is both somewhat conservative -- it will only delete for specific
situations -- and somewhat aggressive -- in that it assumes it's better
to have incorrectly deleted a lockfile that was in active (but slow) use
than to block someone. Since these are actually just config files with
wide validity bands in terms of time, this is probably fine.

Fixed: 107136
Multiply: ffx-e2e

Change-Id: I8076049ded355c40eea87527af5e40e14446ffbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712992
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>