[SDK] Build product bundles for bringup
https://fxrev.dev/712545 made some changes to the build system so that
a product bundle could be generated for bringup images. What that CL
missed was that the bundle was not generated automatically: it had to be
triggered manually, because the target was not included as a dependency
of the build for bringup.

Additionally, the `ffx product verify` tool was just landed, which
included a check to ensure the product bundle contained at least one
entry for FVM images, which is no longer required. This test was causing
the virtual device spec generation to fail for bringup, since bringup
contains no FVMs.

This CL adds in the dependency for bringup product bundles, and removes
the test for FVM images. Combined, these changes result in a successful
run of bringup on the emulator.

Change-Id: I9ee6d4863fae0d6b22a798487b37b6eae42f90d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712867
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>